feat: add retry policy for setting client
