+++ internal/setting/setting.go
@@ -42,6 +42,7 @@ func (s Setting) GetValue() string {
 type Client struct {
 	c           request.Client
 	cred        auth.Credential
+	retryPolicy httpr.RetryPolicy
 	baseURL     string
 	userAgent   string
 	concurrency int
@@ -66,6 +67,7 @@ func NewClient(appConfiguration string, cred auth.Credential, options ...ClientO
 	if c.c == nil {
 		c.c = httpr.NewClient(
 			httpr.WithTimeout(c.timeout),
+			httpr.WithRetryPolicy(c.retryPolicy),
 		)
 	}
 	return c
@@ -207,25 +209,4 @@ func (c Client) getSettings(ctx context.Context, keys []string, options ...Optio
 		settings[sr.key] = sr.setting
 	}
 	return settings, nil
-}
-
-// WithConcurrency sets the concurrency for secret retrieval.
-func WithConcurrency(n int) ClientOption {
-	return func(c *Client) {
-		c.concurrency = n
-	}
-}
-
-// WithTimeout sets timeout for secret retreival.
-func WithTimeout(d time.Duration) ClientOption {
-	return func(c *Client) {
-		c.timeout = d
-	}
-}
-
-// WithLabel sets label on the on a setting request.
-func WithLabel(label string) Option {
-	return func(o *Options) {
-		o.Label = label
-	}
 }
